backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
matplotlib==3.4.2
mysql_connector_repackaged==0.3.1
pandas==1.3.1
